# optimize transpose copy on CPU using fbgemm transpose